% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_anno_parquet.R
\name{available_anno_parquet}
\alias{available_anno_parquet}
\title{use 'paws::s3' to interrogate an NSF Open Storage Network
bucket for parquet serializations of genomic reference annotation}
\usage{
available_anno_parquet()
}
\description{
use 'paws::s3' to interrogate an NSF Open Storage Network
bucket for parquet serializations of genomic reference annotation
}
\examples{
if (requireNamespace("paws")) {
  available_anno_parquet()
}
}
