% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{get_local_table}
\alias{get_local_table}
\title{obtain path to a localized parquet resource}
\usage{
get_local_table(x, con, cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{x}{character(1) stub of use in make_parq_url}

\item{con}{DBI connection}

\item{cache}{BiocFileCache-compliant cache}
}
\description{
obtain path to a localized parquet resource
}
\note{
If cache does not already have a relevant file, one is retrieved from Bioconductor OSN bucket.
}
\examples{
ca = BiocFileCache::BiocFileCache()
con = DBI::dbConnect(duckdb::duckdb())
pa = get_local_table("gene_info.parquet", con, ca)
pa
}
