% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_support.R
\name{make_parq_url}
\alias{make_parq_url}
\title{support for URL generation based on NSF OSN bucket}
\usage{
make_parq_url(stub)
}
\arguments{
\item{stub}{character(1) string like gene2go}
}
\description{
support for URL generation based on NSF OSN bucket
}
