% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_support.R
\name{query_osn}
\alias{query_osn}
\title{produce a tbl with genomic annotation using a remote parquet file}
\usage{
query_osn(con, stub = "gene_info")
}
\arguments{
\item{con}{a DBI connection, typically using duckdb, that can be used
to query remote parquet files}

\item{stub}{character(1) tag identifying an annotation resource}
}
\description{
produce a tbl with genomic annotation using a remote parquet file
}
\examples{
con = DBI::dbConnect(duckdb::duckdb())
query_osn(con, stub="gene_info")
query_osn(con, stub="gene2go")
DBI::dbDisconnect(con)
}
